+++ backend-laravel/app/Http/Controllers/InterestController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\Interest\AddInterestRequest;
+use App\Models\Interest;
 use App\Services\Contracts\InterestServiceInterface;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 
 class InterestController extends Controller
@@ -17,11 +19,14 @@ class InterestController extends Controller
 
     /**
      * Create a new interest.
+     *
+     * @throws AuthorizationException
      */
     public function addInterest(AddInterestRequest $request): JsonResponse
     {
-        $data = $request->validated();
+        $this->authorize('create', Interest::class);
 
+        $data = $request->validated();
         $newInterest = $this->interestService->addInterest($data);
 
         return response()->json([
@@ -32,9 +37,13 @@ class InterestController extends Controller
 
     /**
      * List all interests.
+     *
+     * @throws AuthorizationException
      */
     public function listAllInterests(): JsonResponse
     {
+        $this->authorize('viewAny', Interest::class);
+
         $interests = $this->interestService->getAllInterests();
 
         return response()->json([
@@ -44,9 +53,13 @@ class InterestController extends Controller
 
     /**
      * Delete an existing interest.
+     *
+     * @throws AuthorizationException
      */
     public function deleteInterest(int $interestId): JsonResponse
     {
+        $this->authorize('delete', Interest::class);
+
         $this->interestService->deleteInterest($interestId);
 
         return response()->json([
